Change node address in Endpoint to long

+++ framework/src/main/java/edu/uci/eecs/wukong/framework/model/ComponentMap.java
@@ -21,7 +21,7 @@ public class ComponentMap {
 		this.components.add(component);
 	}
 	
-	public int getPrimaryEndPointNodeId(int componentId) {
+	public long getPrimaryEndPointNodeId(int componentId) {
 		if (componentId < components.size()) {
 			return components.get(componentId).getPrimaryEndPoint().getNodeId();
 		}
